+++ src/components/search_bar/filters/field_value_toggle_filter.js
@@ -63,7 +63,10 @@ export class FieldValueToggleFilter extends Component {
       this.valueChanged(checked);
     };
     return (
-      <EuiFilterButton onClick={onClick} hasActiveFilters={hasActiveFilters}>
+      <EuiFilterButton
+        onClick={onClick}
+        hasActiveFilters={hasActiveFilters}
+        aria-pressed={!!hasActiveFilters}>
         {name}
       </EuiFilterButton>
     );